Refactored user data endpoints
